requirements.txt
numpy~=1.23.5
nltk~=3.7
torch~=1.12.1+cu113
Pillow~=9.3.0
tqdm~=4.64.1
torchvision~=0.13.1+cu113
torchtext~=0.13.1
sklearn~=0.0.post1
scikit-learn~=1.1.3
joblib~=1.2.0
PyYAML~=6.0
matplotlib~=3.6.2
pandas~=1.5.2